fully fix int float and update include statement

+++ app/ACM/Match.cc
@@ -1,4 +1,4 @@
-#include "match.h"
+#include <ACM/Match.h>
 
 Match::Match(Animal* _animal, Client* _client, float _score, QVector<QString> _rules) :
     animal(_animal), client(_client), score(_score), rules(_rules)
@@ -22,7 +22,7 @@ Client* Match::getClient() {
     return client;
 }
 
-int Match::getScore() {
+float Match::getScore() {
     return score;
 }
